WebDevFinal/t_learning/requirements.txt
Django==4.0.0
psycopg2-binary==2.9.3
djangorestframework==3.12.4
pytz==2024.1
django-cors-headers